backend/requirements.txt
# ── Core Framework ─────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# ── Database ──────────────────────────────────────────
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
pgvector>=0.3.6

# ── Auth & Security ───────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
slowapi>=0.1.9

# ── HTTP Clients ──────────────────────────────────────
httpx>=0.28.0
requests>=2.32.0
aiohttp>=3.11.0

# ── AI APIs ───────────────────────────────────────────
anthropic>=0.43.0
openai>=1.58.0

# ── Validation & Config ──────────────────────────────
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1

# ── Redis ─────────────────────────────────────────────
redis>=5.0.0

# ── RAG / Knowledge Base ─────────────────────────────
pypdf>=4.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
tiktoken>=0.7.0

# ── Export ────────────────────────────────────────────
openpyxl>=3.1.0

# ── Email ─────────────────────────────────────────────
aiosmtplib>=3.0.0
jinja2>=3.1.0

# ── Billing ───────────────────────────────────────────
stripe>=8.0.0

# ── Logging & Monitoring ─────────────────────────────
structlog>=24.0.0
sentry-sdk[fastapi]>=2.0.0

# ── Testing ───────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
